Moved hub methods to hub manager class
